+++ Projeto/painel-adm/produtos/excluir.php
@@ -7,7 +7,9 @@ $id = $_POST['id'];
 $query_con = $pdo->query("SELECT * FROM produtos WHERE id = '$id'");
 $res_con = $query_con->fetchAll(pdo::FETCH_ASSOC);
 $imagem = $res_con[0]['foto'];
-unlink('../../img/produtos/' .$imagem);
+if($imagem != 'sem-foto.jpg'){
+    unlink('../../img/produtos/' .$imagem);
+}
 
 $query_con = $pdo->query("DELETE from produtos WHERE id = '$id'");
